Add identification for melee weapons
I've used some tagof() sorcery so that we can effectively overload the
functions, rather than creating IsValidMeleeWeaponId() et al. If anyone
has objections to this, we can duplicate the functions.

This implementation should not break any existing behaviour, because the
MeleeWeaponId versions are only used when the tagType argument is definitely
tagof(MeleeWeaponId). If a variable which is of tag _ is coerced into the
function, tagType == tagof(_) and so the old behaviour applies.